+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   - run standalone, `update` in the package directory calls `./update.ps1`.
 - `README.md` made available via `man about_au`. 
 - Consider global variable `$au_root` when looking for AU packages.
+- Optimization and refactoring.
 - Bugfixes
   - `Update-Packages` exception when `au_GetLatests` returned nothing.
   - `$Latest.Version` remains set to remote version when forcing update.