+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ## NEXT
 
+## 2016.9.14.233253
+
 - `Update-Package`
   - New alias `lsau`.
   - Return an object of type `AUPackage` instead of text.